Split FileStream declarations and definitions
