+++ src/openrct2/FileClassifier.cpp
@@ -10,7 +10,7 @@
 #include "FileClassifier.h"
 
 #include "core/Console.hpp"
-#include "core/FileStream.hpp"
+#include "core/FileStream.h"
 #include "core/Path.hpp"
 #include "rct12/SawyerChunkReader.h"
 #include "scenario/Scenario.h"
