+++ src/openrct2/core/FileIndex.hpp
@@ -13,7 +13,7 @@
 #include "Console.hpp"
 #include "File.h"
 #include "FileScanner.h"
-#include "FileStream.hpp"
+#include "FileStream.h"
 #include "JobPool.hpp"
 #include "Path.hpp"
