+++ src/openrct2/network/ServerList.cpp
@@ -14,7 +14,7 @@
 #    include "../Context.h"
 #    include "../PlatformEnvironment.h"
 #    include "../config/Config.h"
-#    include "../core/FileStream.hpp"
+#    include "../core/FileStream.h"
 #    include "../core/Guard.hpp"
 #    include "../core/Http.h"
 #    include "../core/Json.hpp"
