+++ src/openrct2/title/TitleSequence.cpp
@@ -14,7 +14,7 @@
 #include "../core/Console.hpp"
 #include "../core/File.h"
 #include "../core/FileScanner.h"
-#include "../core/FileStream.hpp"
+#include "../core/FileStream.h"
 #include "../core/Guard.hpp"
 #include "../core/Memory.hpp"
 #include "../core/MemoryStream.h"
